backend/requirements.txt
# Flask Backend Requirements for Barclays RAG Chatbot
# Core Flask dependencies
flask==3.0.3
flask-cors==4.0.1
flask-limiter==3.5.0

# Azure Services
azure-search-documents==11.6.0
azure-core==1.30.2
openai==1.50.2

# Environment & Configuration
python-dotenv==1.0.1

# HTTP & Network
requests==2.32.3
urllib3==2.2.2

# Data Processing
pydantic==2.8.2

# Logging & Monitoring
colorama==0.4.6

# Production Server (optional)
gunicorn==21.2.0
waitress==3.0.0 